+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This file is used to list changes made in each version of the AWS ParallelCluste
 - Run daemons as cluster admin user (not root).
 - Add explicit assignment of names, uids, gids for slurm, munge and dcvextauth users.
 - Remove packer.
+- Restrict access to IMDS to root and cluster admin users, only.
 
 
 2.x.x